+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 4. **Compatibility Warnings**:
    1. move `async_logger`, `AsyncQueueListener`, `LogHelper`, `RotatingFileWriter`, and `SizedTimedRotatingFileHandler` from `morebuiltins.funcs` to `morebuiltins.logs`.
 5. add `ContextFilter` for logging, to add context variables to log records.
+6. add `queue` arg to `morebuiltins.logs.LogHelper.bind_handler`, to enable async logging with `AsyncQueueListener`. (like loguru's queue=True)
 
 ### 1.3.1 (2025-06-19)
 1. add `check_recursion` to `morebuiltins.funcs`, to check if a function is recursive.